Add accessibility; create AR hook
