+++ src/modules/service/components/ServicePointDetailsImage.tsx
@@ -1,35 +1,13 @@
-import {useEffect, useMemo, useState} from 'react'
-import {Image} from 'react-native'
 import {LazyImage} from '@/components/ui/media/LazyImage'
-import {getClosestAspectRatio} from '@/modules/service/utils/getClosestAspectRatio'
+import {useDynamicImageAspectRatio} from '@/hooks/useDynamicImageAspectRatio'
 
 export const ServicePointDetailsImage = ({uri}: {uri: string}) => {
-  const [dimensions, setDimensions] = useState<{
-    height: number
-    width: number
-  } | null>(null)
-
-  useEffect(() => {
-    Image.getSize(
-      uri,
-      (width, height) => setDimensions({width, height}),
-      () => setDimensions(null),
-    )
-  }, [uri])
-
-  const aspectRatio = useMemo(
-    () =>
-      dimensions
-        ? getClosestAspectRatio(dimensions.width, dimensions.height)
-        : undefined,
-    [dimensions],
-  )
+  const aspectRatio = useDynamicImageAspectRatio(uri)
 
   return (
     <LazyImage
       aspectRatio={aspectRatio}
       fallbackInheritsAspectRatio={false}
-      openInImageViewer
       source={{uri}}
       testID="ServicePointDetailsPropertiesImage"
     />
